+++ foundation_struct/init.lua
@@ -2,7 +2,7 @@
 -- Foundation Struct
 --
 -- Adds wrapper classes for some basic structures
-local mod = foundation.new_module("foundation_struct", "1.5.0")
+local mod = foundation.new_module("foundation_struct", "1.6.0")
 
 mod:require("struct/list.lua")
 mod:require("struct/linked_list.lua")