+++ foundation_struct/CHANGELOG.md
@@ -1,3 +1,10 @@
+# 1.6.0
+
+* Added `List#bsearch/1`
+* Added `List#bsearch_by/1`
+* Fixed bad implementation of `List#to_linked_list/0`
+* Fixed bad implementation of `LinkedList#to_list/0`
+
 # 1.5.0
 
 * Added `List#reduce_while/2`